pyproject.toml
[project]
name = "rag-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.4",
    "faiss-cpu>=1.12.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "pymupdf>=1.26.6",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "typesense>=1.3.0",
]
